+++ src/id_translation/utils/__init__.py
@@ -1,7 +1,7 @@
 """Various supporting functions and classes."""
 
-# from .._utils.emit_warning import add_skip_file_prefix
-# # TODO(2.0.0): Expose + document + integrate with id-translation-project.
-# __all__ = [
-#    "add_skip_file_prefix",
-# ]
+from .._utils.emit_warning import add_skip_file_prefix
+
+__all__ = [
+    "add_skip_file_prefix",
+]