+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `Format.placeholder_attributes` property.
 - Add `FetchInstruction.placeholder_attributes` property + `AbstractFetcher` handling.
 - Add SQLAlchemy ORM fetcher example.
+- Add the `add_skip_file_prefix()` function to control warning stack levels.
 
 ### Changed
 - Links in warnings and exceptions now to the current version in the docs (instead of the _'stable'_ version).